refactor: add custom hook for column resizing
