+++ packages/elements-react/lib/components/Lens/hooks/index.ts
@@ -1,3 +1,6 @@
 export { useFetchConfig } from "./src/useFetchConfig";
 export { useTableConfig } from "./src/useTableConfig";
 export { useResizeObserver } from "./src/useResizeObserver";
+export { useColumnSizing } from "./src/useColumnSizing";
+
+export type { ResizeObserverCallback } from "./src/useResizeObserver";
